+++ test/parse_test.dart
@@ -509,6 +509,28 @@ void main() {
         var results = parser.parse(['--b=1']);
         expect(results['a'], '1');
       });
+
+      group('mandatory', () {
+        test('throw if no args', () {
+          var parser = ArgParser();
+          parser.addOption('username', mandatory: true);
+          throwsFormat(parser, []);
+        });
+
+        test('throw if no mandatory args', () {
+          var parser = ArgParser();
+          parser.addOption('test');
+          parser.addOption('username', mandatory: true);
+          throwsFormat(parser, ['--test', 'test']);
+        });
+
+        test('parse successfully', () {
+          var parser = ArgParser();
+          parser.addOption('test', mandatory: true);
+          var results = parser.parse(['--test', 'test']);
+          expect(results['test'], equals('test'));
+        });
+      });
     });
 
     group('remaining args', () {