+++ test/usage_test.dart
@@ -5,6 +5,8 @@
 import 'package:args/args.dart';
 import 'package:test/test.dart';
 
+import 'test_utils.dart';
+
 void main() {
   group('ArgParser.usage', () {
     test('negatable flags show "no-" in title', () {
@@ -402,6 +404,22 @@ void main() {
           --[no-]small2         a
           --[no-]small3         a
           ''');
+    });
+
+    test('display "mandatory" after a mandatory option', () {
+      var parser = ArgParser();
+      parser.addOption('test', mandatory: true);
+      validateUsage(parser, '''
+        --test (mandatory)    
+        ''');
+    });
+
+    test('throw argument error if option is mandatory with a default value', () {
+      var parser = ArgParser();
+      expect(
+        () => parser.addOption('test', mandatory: true, defaultsTo: 'test'),
+        throwsArgumentError
+      );
     });
 
     group('separators', () {