+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.1.0-dev
 
+* Add a `mandatory` argument to require the presence of an option.
 * Add `aliases` named argument to `addFlag`, `addOption`, and `addMultiOption`,
   as well as a public `findByNameOrAlias` method on `ArgParser`. This allows
   you to provide aliases for an argument name, which eases the transition from